requirements.txt
Flask
Flask-Cors
flask-session
PyMySQL
mysql-connector-python
Pillow